implement proper deleting of interface

+++ src/lib/y2network/widgets/delete_interface.rb
@@ -19,6 +19,7 @@
 
 require "yast"
 require "cwm/common_widgets"
+require "yast2/popup"
 
 Yast.import "Label"
 Yast.import "Lan"
@@ -51,9 +52,36 @@ module Y2Network
           end
         end
 
+        others = all_modify(config, connection_config)
+        if !others.empty?
+          delete, modify = others.partition { |c| c.type.vlan? }
+          message = format(_("Device you select has been used in other devices.<br>" \
+            "When deleted these devices will be modified<ul>%s</ul><br>" \
+            "and these devices deleted: <ul>%s</ul><br>" \
+            "Really delete?"),
+            modify.map { |m| "<li>#{m.name}</li>" }.join("\n"),
+            delete.map { |m| "<li>#{m.name}</li>" }.join("\n")
+          )
+
+          if Yast2::Popup.show(message, richtext: :yes, buttons: :yes_no, headline: :warning) == :no
+            return nil
+          end
+        end
+
         config.delete_interface(@table.value)
 
         :redraw
+      end
+
+    private
+
+      # @return [Array]
+      def all_modify(config, connection_config)
+        all = config.connections_to_modify(connection_config).to_a
+        vlans = all.select { |c| c.type.vlan? }
+        vlans.each_with_object(all) { |c, a| a.concat(all_modify(config, c)) }
+
+        all
       end
     end
   end
